소셜 로그인 ,회원 탈퇴,상세 조회, display추가, 서비스 분리
