+++ src/main/java/com/imfine/ngs/user/service/ProfileService.java
@@ -1,0 +1,37 @@
+package com.imfine.ngs.user.service;
+
+import com.imfine.ngs.user.dto.response.UserProfileDto;
+import com.imfine.ngs.user.entity.User;
+import com.imfine.ngs.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ProfileService {
+
+    private final UserRepository userRepository;
+
+    public void updateNickname(String email, String newNickname) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디"));
+
+        user.updateNickname(newNickname);
+        userRepository.save(user);
+    }
+
+    public UserProfileDto getUserProfile(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디"));
+
+        return new UserProfileDto(user.getEmail(), user.getNickname());
+    }
+
+    public void deleteUser(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 아이디"));
+
+        userRepository.delete(user);
+    }
+}
+
